Phase 1 — Enforce MySQL as the platform database (migration tooling + decimal money) (#37)
Makes MySQL the required database for staging/production and DB integration tests, with no silent
SQLite fallback. The backend is already provider-agnostic (Data/Db.cs translates SQLite-dialect SQL
to MySQL); this phase enforces, hardens and tools the switch — no application rewrite.

- Production guard (Program.cs): in ASPNETCORE_ENVIRONMENT=Production the preflight fails unless
  DB_PROVIDER=mysql and MySQL connection settings are present — SQLite is refused in production.
- Connection robustness (Db.cs): configurable pooling (MYSQL_POOL_MIN/MAX), command timeout,
  connect timeout, and a boot-time connect retry with backoff (managed MySQL can be briefly
  unavailable at deploy/failover) — still no fallback to SQLite.
- Money precision (schema.mysql.sql): the 11 money columns are DECIMAL(12,2) instead of DOUBLE
  (payments, memberships, pricing_rules, certifications, discount_codes, code_redemptions);
  percentages / scores / CPD hours stay DOUBLE. Verified 149.999 -> 150.00 round-trip.
- Migration + reconciliation tool (backend/tools/migrate_sqlite_to_mysql.py): copies all tables in
  dependency-safe order (FK checks off during load), preserves primary keys and password hashes,
  fixes AUTO_INCREMENT, then reconciles row counts, financial totals, FK integrity and uniqueness,
  emitting a JSON report. Idempotent per table; source is never modified.
- Backup helper (backend/tools/mysql_backup.sh): consistent gzipped mysqldump --single-transaction
  with retention.
- Deployment (render.yaml): switched to external managed MySQL (DB_PROVIDER=mysql, MYSQL_* secrets,
  SSL required); the persistent disk now holds uploaded files only.
- Docs (docs/MYSQL_MIGRATION.md): server requirements, env templates (dev/test/staging/prod),
  migration, cutover, backup/restore/PITR, rollback.

Verified live against a local MariaDB 10.11: app boots + serves on MySQL; a seeded SQLite (63 tables,
14,068 rows incl. sample users/payments/memberships) migrated with a clean reconciliation (exact row
counts, all financial totals matched, 0 FK orphans, 0 dup uniques); the bcrypt password hash was
preserved byte-identical and the migrated owner logs in. Adversarial suites on MySQL: integration
148/148, founding 46/46, honorary 19/19, honorary-application 20/20.

+++ backend/schema.mysql.sql
@@ -40,7 +40,7 @@ CREATE TABLE IF NOT EXISTS student_profiles (
 CREATE TABLE IF NOT EXISTS pricing_rules (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
   currency TEXT DEFAULT 'USD', product_type TEXT,           -- membership | exam | renewal | retake
-  standard_price DOUBLE, default_discount_percentage DOUBLE DEFAULT 0,
+  standard_price DECIMAL(12,2), default_discount_percentage DOUBLE DEFAULT 0,
   active BIGINT DEFAULT 1, start_date TEXT, end_date TEXT,
   created_at TEXT DEFAULT (DATE_FORMAT(UTC_TIMESTAMP(),'%Y-%m-%d %H:%i:%s')), updated_at TEXT DEFAULT (DATE_FORMAT(UTC_TIMESTAMP(),'%Y-%m-%d %H:%i:%s'))
 );
@@ -48,7 +48,7 @@ CREATE TABLE IF NOT EXISTS discount_codes (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
   code VARCHAR(500) UNIQUE NOT NULL,
   discount_type TEXT,                        -- percentage | fixed
-  discount_value DOUBLE,
+  discount_value DECIMAL(12,2),
   applies_to TEXT DEFAULT 'all',             -- membership | exam | all
   start_date TEXT, end_date TEXT,
   max_uses BIGINT, used_count BIGINT DEFAULT 0,
@@ -72,8 +72,8 @@ CREATE TABLE IF NOT EXISTS discount_codes (
 CREATE TABLE IF NOT EXISTS payments (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
   user_id BIGINT, enrollment_session_id BIGINT,
-  product_type TEXT, standard_amount DOUBLE, default_discount_amount DOUBLE,
-  discount_code TEXT, discount_code_amount DOUBLE, final_amount DOUBLE, currency TEXT DEFAULT 'USD',
+  product_type TEXT, standard_amount DECIMAL(12,2), default_discount_amount DECIMAL(12,2),
+  discount_code TEXT, discount_code_amount DECIMAL(12,2), final_amount DECIMAL(12,2), currency TEXT DEFAULT 'USD',
   payment_provider TEXT, provider_payment_id TEXT,
   payment_status TEXT,                       -- paid | failed | refunded
   payment_date TEXT, invoice_url TEXT, receipt_url TEXT,
@@ -85,8 +85,8 @@ CREATE TABLE IF NOT EXISTS memberships (
   user_id BIGINT NOT NULL,
   membership_type TEXT, status TEXT DEFAULT 'inactive',   -- inactive | active | expired
   start_date TEXT, expiry_date TEXT,
-  renewal_fee DOUBLE DEFAULT 99, renewal_cycle TEXT DEFAULT '3 years',
-  amount_paid DOUBLE, currency TEXT DEFAULT 'USD'
+  renewal_fee DECIMAL(12,2) DEFAULT 99, renewal_cycle TEXT DEFAULT '3 years',
+  amount_paid DECIMAL(12,2), currency TEXT DEFAULT 'USD'
 );
 CREATE TABLE IF NOT EXISTS login_tokens (
   id BIGINT PRIMARY KEY AUTO_INCREMENT, user_id BIGINT NOT NULL,
@@ -345,7 +345,7 @@ CREATE TABLE IF NOT EXISTS certifications (
   pass_mark_pct DOUBLE,                        -- NULL → global exam_pass_mark_pct
   duration_minutes BIGINT,                  -- NULL → global exam_duration_minutes
   expiry_years BIGINT DEFAULT 3,
-  exam_price DOUBLE,                           -- NULL → pricing_rules product_type='exam'
+  exam_price DECIMAL(12,2),                           -- NULL → pricing_rules product_type='exam'
   active BIGINT DEFAULT 1,
   sort_order BIGINT DEFAULT 0,
   created_at TEXT DEFAULT (DATE_FORMAT(UTC_TIMESTAMP(),'%Y-%m-%d %H:%i:%s')), updated_at TEXT DEFAULT (DATE_FORMAT(UTC_TIMESTAMP(),'%Y-%m-%d %H:%i:%s'))
@@ -717,7 +717,7 @@ CREATE TABLE IF NOT EXISTS code_redemptions (
   id BIGINT PRIMARY KEY AUTO_INCREMENT,
   code_id BIGINT, code TEXT, user_id BIGINT, email TEXT,
   payment_id BIGINT UNIQUE,                 -- one redemption per payment (idempotent)
-  product_type TEXT, amount_before DOUBLE, discount_amount DOUBLE,
+  product_type TEXT, amount_before DECIMAL(12,2), discount_amount DECIMAL(12,2),
   redeemed_at TEXT DEFAULT (DATE_FORMAT(UTC_TIMESTAMP(),'%Y-%m-%d %H:%i:%s'))
 );
 CREATE INDEX IF NOT EXISTS ix_redemptions_code ON code_redemptions(code_id);
